utils: Move the xmlinc tool into libs.
